Time: 70 ms (9.84%), Space: 11.7 MB (35.52%) - LeetHub

+++ 1759-count-number-of-homogenous-substrings/1759-count-number-of-homogenous-substrings.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    long long sum(int n){
+    int sum(int n){
         long long res=0;
         while(n!=0){
             res+=n;
